stop writing temporary .json dynmat files
the 'payoff' of the other recent commits

+++ src/python/rsp2/internals/convert/__init__.py
@@ -31,6 +31,8 @@ def _main(
         input_is_stdin: bool,
         output_is_stdout: bool,
 ):
+    keep = keep or output_is_stdout
+
     if os.path.exists(output):
         if os.path.samefile(output, input):
             return
@@ -50,7 +52,7 @@ def _main(
         info("Internal error; output does not match input")
         sys.exit(1)
 
-    if not keep:
+    if not input_is_stdin and not keep:
         os.unlink(input)
 
 def _main_from_cli(
@@ -60,8 +62,7 @@ def _main_from_cli(
 
     p.add_argument('INPUT')
     p.add_argument('--output', '-o', required=True)
-    p.add_argument('--stdin', action='store_true', help='read from stdin, using the file extension of INPUT to determine the format')
-
+    p.add_argument('--stdin', action='store_true', help='read from stdin, using the file extension of INPUT to determine the format. Implies --keep.')
     p.add_argument('--stdout', action='store_true', help='write to stdout, using the file extension of --output to determine the format')
     p.add_argument('--keep', action='store_true')
     args = p.parse_args()
@@ -69,27 +70,47 @@ def _main_from_cli(
     _main(
         hooks=hooks,
         keep=args.keep,
-        output=args.output,
         input=args.INPUT,
+        output=args.output,
         input_is_stdin=args.stdin,
         output_is_stdout=args.stdout,
     )
 
-def _main_from_rust(
+def _read_from_rust(
         hooks: DwimHooks,
 ):
-    d = json.load(sys.stdin)
-    keep = d.pop('keep')
-    output = d.pop('output')
-    input = d.pop('input')
-    assert not d
+    p = argparse.ArgumentParser()
+    p.add_argument('INPUT')
+    args = p.parse_args()
 
+    # Receive a unit type
+    assert json.load(sys.stdin) in [None, []]
+
+    # Send the matrix
     _main(
         hooks=hooks,
-        keep=keep,
-        output=output,
-        input=input,
+        keep=False,
+        input=args.INPUT,
+        output="<stdout>.json",
         input_is_stdin=False,
+        output_is_stdout=True,
+    )
+
+def _write_from_rust(
+        hooks: DwimHooks,
+):
+    p = argparse.ArgumentParser()
+    p.add_argument('OUTPUT')
+    args = p.parse_args()
+
+    # Receive the matrix and write it
+    _main(
+        hooks=hooks,
+        keep=False,
+        input="<stdin>.json",
+        output=args.OUTPUT,
+        input_is_stdin=True,
         output_is_stdout=False,
     )
+    # Send a unit type
     print("null")
